# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Origin Agent Platform Product API
  version: v1.0-alpha
  description: >
    Product-level API contract for Origin Agent Platform. The product owns this
    contract; server-side services implement it; PloyKit modules consume it.
...
servers:
  - url: https://api.origin.example
    description: Production API endpoint.
  - url: http://127.0.0.1:8080
    description: Local implementation endpoint.
  - url: http://127.0.0.1:4010
    description: Local contract/mock endpoint.
security:
  - serviceToken: []
    originSignature: []
    originTimestamp: []
    originClaims: []
...
paths:
...
  /v1/runs:
    get:
      tags: [Runs]
      operationId: listRuns
      summary: List agent runs.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: agent_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Run list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunListResponse"
    post:
      tags: [Runs]
      operationId: startRun
      summary: Start an agent run.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunStartRequest"
      responses:
        "202":
          description: Run accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResponse"
...
components:
  securitySchemes:
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: service-token
    originSignature:
      type: apiKey
      in: header
      name: x-origin-agentops-signature
    originTimestamp:
      type: apiKey
      in: header
      name: x-origin-agentops-timestamp
    originClaims:
      type: apiKey
      in: header
      name: x-origin-agentops-claims
  parameters:
...
    IdempotencyKey:
      name: idempotency-key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
...
  schemas:
    Meta:
      type: object
      required: [request_id, trace_id]
      properties:
        request_id:
          type: string
        trace_id:
          type: string
        workspace_id:
          type: string
        tenant_id:
          type: string
      additionalProperties: true
    ListMeta:
      allOf:
        - $ref: "#/components/schemas/Meta"
        - type: object
          properties:
            next_cursor:
              type: string
              nullable: true
    ErrorEnvelope:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        meta:
          $ref: "#/components/schemas/Meta"
...
    Run:
      type: object
      required: [run_id, agent_id, status, objective, created_at]
      properties:
        run_id:
          type: string
        agent_id:
          type: string
        agent_name:
          type: string
        agent_version:
          type: string
        task_id:
          type: string
        trace_id:
          type: string
        status:
          type: string
          enum: [queued, running, waiting_input, waiting_approval, approval_required, succeeded, failed, cancelled]
        trigger:
          type: string
        environment:
          type: string
        user:
          type: string
        objective:
          type: string
        tool_calls:
          type: integer
        tokens:
          type: integer
        duration_ms:
          type: integer
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
        artifact_refs:
          type: array
          items:
            type: string
        result:
          type: object
          additionalProperties: true
        error:
          $ref: "#/components/schemas/ErrorEnvelope"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
    RunStartRequest:
      type: object
      required: [agent_id, objective]
      properties:
        agent_id:
          type: string
        objective:
          type: string
        input:
          type: object
          additionalProperties: true
        trace_id:
          type: string
        trigger:
          type: string
        environment:
          type: string
        metadata:
          type: object
          additionalProperties: true
...
    RunResponse:
      type: object
      required: [run, meta]
      properties:
        run:
          $ref: "#/components/schemas/Run"
        meta:
          $ref: "#/components/schemas/Meta"
    RunListResponse:
      type: object
      required: [runs, meta]
      properties:
        runs:
          type: array
          items:
            $ref: "#/components/schemas/Run"
        meta:
          $ref: "#/components/schemas/ListMeta"
...
    TimelineEvent:
      type: object
      required: [event_id, title, occurred_at]
      properties:
        event_id:
          type: string
        title:
          type: string
        status:
          type: string
        category:
          type: string
        occurred_at:
          type: string
          format: date-time
        duration_ms:
          type: integer
        actor:
          type: string
        payload:
          type: object
          additionalProperties: true
...